refactor: separate chart component
