+++ Client/.editorconfig
@@ -1,0 +1,9 @@
+[*]
+charset = utf-8
+indent_style = space
+indent_size = 2
+trim_trailing_whitespace = true
+insert_final_newline = true
+
+[*.md]
+max_line_length = 150
