+++ Client/.vscode/settings.json
@@ -7,5 +7,6 @@
         "**/package-lock.json": true
     },
     "editor.bracketPairColorization.enabled": true,
-    "editor.guides.bracketPairs": "active"
+    "editor.guides.bracketPairs": "active",
+    "editor.formatOnSave": true
 }
