+++ Client/src/app/app.component.scss
@@ -1,72 +1,33 @@
 @import "../styles-theme";
 
-// PAGE CONTAINERS
+// TOOLBAR
+.toolbar {
+  a {
+    text-decoration: none;
+    color: $color-primary-light;
+  }
 
-.body-container {
-    margin: 16px;
-    padding-bottom: 16px;
+  font-family: Google Sans, Roboto, Arial, Helvetica, sans-serif;
+
+  .filler {
+    flex: 1 1 auto;
+    text-align: center;
+  }
 }
 
+// PAGE CONTAINERS
 .page-intro {
-    max-width: 950px;
-    text-align: center;
-    margin: auto;
-    margin-bottom: 1em;
+  max-width: 950px;
+  text-align: center;
+  margin: auto;
+  margin-bottom: 1em;
 }
 
 .badge-container {
-    text-align: center;
-    .badge {
-        margin-top: 0.25em;
-        padding-left: 0.25em;
-        padding-right: 0.25em;
-    }
-}
-
-.chart-container {
-    padding: 0;
-    margin: 0;
-    border-style: none;
-    border-width: 0;
-}
-
-// TOOLBAR
-.toolbar {
-    a {
-        text-decoration: none;
-        color: $color-primary-light;
-    }
-
-    font-family: Google Sans, Roboto, Arial, Helvetica, sans-serif;
-
-    .filler {
-        flex: 1 1 auto;
-        text-align: center;
-    }
-}
-
-// CONTROLS
-
-mat-chip-list {
-    margin: 16px;
-    display: flex;
-    justify-content: center;
-}
-
-mat-radio-group {
-    margin: 16px;
-}
-
-.radio-button {
-    margin: 5px;
-}
-
-.button-row {
-    margin-top: 25px;
-    margin-bottom: 25px;
-}
-
-.button-row > button {
-    margin-left: 4px;
-    margin-right: 4px;
+  text-align: center;
+  .badge {
+    margin-top: 0.25em;
+    padding-left: 0.25em;
+    padding-right: 0.25em;
+  }
 }
