+++ Client/src/app/chart/chart.component.scss
@@ -1,0 +1,6 @@
+.chart-container {
+  padding: 0;
+  margin: 0;
+  border-style: none;
+  border-width: 0;
+}
