+++ Client/src/styles.scss
@@ -1,6 +1,7 @@
 @use "@angular/material" as mat;
 @import "styles-theme";
 
+// BASE STYLES
 html,
 body {
     height: 100%;
@@ -28,4 +29,35 @@ body {
     overflow: hidden;
     text-overflow: ellipsis;
     white-space: nowrap;
+}
+
+.body-container {
+    margin: 16px;
+    padding-bottom: 16px;
+}
+
+// ANGULAR MATERIAL OVERRIDES
+mat-chip-list {
+    margin: 16px;
+    display: flex;
+    justify-content: center;
+}
+
+mat-radio-group {
+    margin: 16px;
+}
+
+// COMMON CONTROLS
+.radio-button {
+    margin: 5px;
+}
+
+.button-row {
+    margin-top: 25px;
+    margin-bottom: 25px;
+}
+
+.button-row > button {
+    margin-left: 4px;
+    margin-right: 4px;
 }
